Update Main
I've edited the main html page so there are links to Bootstrap and JavaScript, an external CSS file that we can customise (I haven't made this yet but will soon) and a title for the tab itself. I've also set the language for the page (en) to make it more easily accessible to screenreaders.