Appearance fiddling (#3)
* Extraneous Code

Removed a few unnecessary bits of commented out code.

* Image playing

Started fiddling with hover animations for the down arrow button.

* Animated Arrow

I added a hover effect to the arrow. It turns pink now!

* Contact Form

Created a contact form with 123formbuilder to add to the contact section of the website.

* Who Are We

Updated formatting so the cards are the same size.

* Heading Stripes

Added in some formatting to the About headings.

* Colour Coordination

Settled on a colour for the About page strips.

* About Formatting

Added more colour and played around with the About heading.

* Minor Updates

Updated text size and info.