Pre-presentation touches (#4)
* Accessible Link

Attempted to add class that reintroduces the hyperlink underlines to make links more accessible. It didn't work.

* Up Arrow

Linked arrow at the bottom of the page to the chart.

* Up Arrow 2

Added up arrow to the bottom of the About page.

* Card Margin

Added more of a margin around the bio deck of cards.

* Fiddling

What the title says.